Fix some bugs, add examples
* Change base_ in BalancedEncoder and BalancedFractionalEncoder to be uint64_t
* Add more code examples and much more details/comments
* Fix bug in SimulationEvaluator::add that causes heap corruption in some
* cases
* Change multiply_many/multiply_norelin_many to make a copy of vector
* inputs instead of using references
* Add utilities.h to seal.h
* Fix a bug in BalancedFractionalEncoder::encode(...) that causes it to
* yield incorrect results in some cases
* Create BIN_DIR automatically by SEAL makefile if it doesn't exist

+++ SEAL/chooser.h
@@ -785,7 +785,7 @@ namespace seal
         @param[in] base The base (default value is 3)
         @throws std::invalid_argument if base is not an odd integer and at least 3
         */
-        ChooserEncoder(int base = 3);
+        ChooserEncoder(uint64_t base = 3);
 
         /**
         Encodes a number (represented by std::uint64_t) into a ChooserPoly object. This is done by first
@@ -887,7 +887,7 @@ namespace seal
         /**
         Returns the base used for encoding.
         */
-        int base() const
+        uint64_t base() const
         {
             return encoder_.base();
         }
